optimized get_visible and added frame buf
get_visible use to check each pixel to see if it was visible. now it just calculates the upper left pixel of the screen based off the camera position and the coords of the upper-left-most chunk loaded into memory then draws each pixel based off the screen size from the upper left pixel.
also made it so the screen doesnt reallocate a new buffer on each frame and instead uses the same buf every time